refactor: combine charge + stream into single server per framework
Each example now has one server with both endpoints:
- GET /api/fortune — pay-per-request (charge)
- GET /api/chat — streaming per-token (stream)

Amp-Thread-ID: https://ampcode.com/threads/T-019c40d1-78b5-7217-90d5-ff3991b8e613
