+++ examples/express/package.json
@@ -4,7 +4,6 @@
   "private": true,
   "scripts": {
     "dev": "tsx src/server.ts",
-    "stream": "tsx src/stream.ts",
     "client": "tsx src/client.ts"
   },
   "dependencies": {
